Rename "env" store to "dotenv"
