+++ decrypt/decrypt.go
@@ -13,7 +13,7 @@ import (
 	"go.mozilla.org/sops/aes"
 	sopsjson "go.mozilla.org/sops/stores/json"
 	sopsyaml "go.mozilla.org/sops/stores/yaml"
-	sopsenv "go.mozilla.org/sops/stores/env"
+	sopsdotenv "go.mozilla.org/sops/stores/dotenv"
 )
 
 // File is a wrapper around Data that reads a local encrypted
@@ -29,7 +29,7 @@ func File(path, format string) (cleartext []byte, err error) {
 
 // Data is a helper that takes encrypted data and a format string,
 // decrypts the data and returns its cleartext in an []byte.
-// The format string can be `json`, `yaml` or `binary`.
+// The format string can be `json`, `yaml`, `dotenv` or `binary`.
 // If the format string is empty, binary format is assumed.
 func Data(data []byte, format string) (cleartext []byte, err error) {
 	// Initialize a Sops JSON store
@@ -39,8 +39,8 @@ func Data(data []byte, format string) (cleartext []byte, err error) {
 		store = &sopsjson.Store{}
 	case "yaml":
 		store = &sopsyaml.Store{}
-	case "env":
-		store = &sopsenv.Store{}
+	case "dotenv":
+		store = &sopsdotenv.Store{}
 	default:
 		store = &sopsjson.BinaryStore{}
 	}
